move editable and stringonenter
