+++ src/PlutoExtras.jl
@@ -7,8 +7,9 @@ using AbstractPlutoDingetjes
 using Reexport
 @reexport using PlutoUI
 
-module EditableModule include("editable.jl") end # exports: Editable 
-module StringOnEnterModule include("stringonenter.jl") end # exports: StringOnEnter
+export Editable, StringOnEnter # from basic_widgets.jl
+
+include("basic_widgets.jl")
 include("latex_equations.jl") # exports the following: texeq, eqref, initialize_eqref, @texeq_str
 module ExtendedToc include("extended_toc.jl") end
 import .ExtendedToc: ExtendedTableOfContents, show_output_when_hidden
