+++ test/notebooks/basic_widgets.jl
@@ -1,0 +1,254 @@
+### A Pluto.jl notebook ###
+# v0.19.27
+
+using Markdown
+using InteractiveUtils
+
+# This Pluto notebook uses @bind for interactivity. When running this notebook outside of Pluto, the following 'mock version' of @bind gives bound variables a default value (instead of an error).
+macro bind(def, element)
+    quote
+        local iv = try Base.loaded_modules[Base.PkgId(Base.UUID("6e696c72-6542-2067-7265-42206c756150"), "AbstractPlutoDingetjes")].Bonds.initial_value catch; b -> missing; end
+        local el = $(esc(element))
+        global $(esc(def)) = Core.applicable(Base.get, el) ? Base.get(el) : iv(el)
+        el
+    end
+end
+
+# ╔═╡ cfb4d354-e0b2-4a04-a559-4fb88df33954
+using PlutoDevMacros
+
+# ╔═╡ 57c51c71-fd8d-440d-8262-9cccd1617c08
+@frompackage "../.." begin
+	using ^
+end
+
+# ╔═╡ d9762fc1-fa2c-4315-a360-cc1cd9d70055
+md"""
+This is an example of `StringOnEnter`: $(@bind text StringOnEnter("ciao"))
+"""
+
+# ╔═╡ cbeff2f3-3ffe-4aaa-8d4c-43c44ee6d59f
+text === "ciao" || error("Something went wrong")
+
+# ╔═╡ ecfdee44-3e6c-4e7e-a039-1f7d05a875f8
+md"""
+This is a number: $(@bind num Editable(3))
+"""
+
+# ╔═╡ 9a90e5be-6c38-440d-a6d6-99dc25be5727
+num === 3 || error("something went wrong")
+
+# ╔═╡ ca6e57a2-667f-4a77-a1ea-5f42192056d4
+@bind(
+	number_editable,
+	Editable(0.5123234; prefix = "Before: ", suffix = " After!", format = ".2f")
+)
+
+# ╔═╡ 81c03bad-aed6-4b45-b3b5-edbba30c0bdb
+number_editable === 0.5123234 || error("something went wrong")
+
+# ╔═╡ ac83e4d6-1637-4646-865c-7d21ce010374
+@bind(
+	bool_editable,
+	Editable(true)
+)
+
+# ╔═╡ 09458943-15a4-4251-82a8-e43299db53b5
+bool_editable || error("something went wrong")
+
+# ╔═╡ d171e8f9-c939-4747-a748-5568ae4a4064
+num |> typeof
+
+# ╔═╡ 9dfb5236-9475-4bf6-990a-19f8f5519003
+md"""
+This has also a unit $(@bind unitnum Editable(3.0;suffix=" dB"))
+"""
+
+# ╔═╡ 245fad1a-2f1e-4776-b048-6873e8c33f3b
+unitnum === 3.0 || error("Something went wrong")
+
+# ╔═╡ 456a3579-3c33-496c-bbf2-6e6e0d0ff102
+bool_bond = @bind bool_val Editable(true)
+
+# ╔═╡ 8f7d7dac-d5e6-43f4-88e6-21179c78c3ef
+bool_val
+
+# ╔═╡ 00000000-0000-0000-0000-000000000001
+PLUTO_PROJECT_TOML_CONTENTS = """
+[deps]
+PlutoDevMacros = "a0499f29-c39b-4c5c-807c-88074221b949"
+"""
+
+# ╔═╡ 00000000-0000-0000-0000-000000000002
+PLUTO_MANIFEST_TOML_CONTENTS = """
+# This file is machine-generated - editing it directly is not advised
+
+[[ArgTools]]
+uuid = "0dad84c5-d112-42e6-8d28-ef12dabb789f"
+version = "1.1.1"
+
+[[Artifacts]]
+uuid = "56f22d72-fd6d-98f1-02f0-08ddc0907c33"
+
+[[Base64]]
+uuid = "2a0f44e3-6c83-55bd-87e4-b1978d98bd5f"
+
+[[Dates]]
+deps = ["Printf"]
+uuid = "ade2ca70-3891-5945-98fb-dc099432e06a"
+
+[[Downloads]]
+deps = ["ArgTools", "FileWatching", "LibCURL", "NetworkOptions"]
+uuid = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
+version = "1.6.0"
+
+[[FileWatching]]
+uuid = "7b1f6079-737a-58dc-b8bc-7a2ca5c1b5ee"
+
+[[HypertextLiteral]]
+deps = ["Tricks"]
+git-tree-sha1 = "c47c5fa4c5308f27ccaac35504858d8914e102f9"
+uuid = "ac1192a8-f4b3-4bfe-ba22-af5b92cd3ab2"
+version = "0.9.4"
+
+[[InteractiveUtils]]
+deps = ["Markdown"]
+uuid = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
+
+[[LibCURL]]
+deps = ["LibCURL_jll", "MozillaCACerts_jll"]
+uuid = "b27032c2-a3e7-50c8-80cd-2d36dbcbfd21"
+version = "0.6.4"
+
+[[LibCURL_jll]]
+deps = ["Artifacts", "LibSSH2_jll", "Libdl", "MbedTLS_jll", "Zlib_jll", "nghttp2_jll"]
+uuid = "deac9b47-8bc7-5906-a0fe-35ac56dc84c0"
+version = "8.0.1+1"
+
+[[LibGit2]]
+deps = ["Base64", "NetworkOptions", "Printf", "SHA"]
+uuid = "76f85450-5226-5b5a-8eaa-529ad045b433"
+
+[[LibSSH2_jll]]
+deps = ["Artifacts", "Libdl", "MbedTLS_jll"]
+uuid = "29816b5a-b9ab-546f-933c-edad1886dfa8"
+version = "1.11.0+1"
+
+[[Libdl]]
+uuid = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
+
+[[Logging]]
+uuid = "56ddb016-857b-54e1-b83d-db4d58db5568"
+
+[[MacroTools]]
+deps = ["Markdown", "Random"]
+git-tree-sha1 = "9ee1618cbf5240e6d4e0371d6f24065083f60c48"
+uuid = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
+version = "0.5.11"
+
+[[Markdown]]
+deps = ["Base64"]
+uuid = "d6f4376e-aef5-505a-96c1-9c027394607a"
+
+[[MbedTLS_jll]]
+deps = ["Artifacts", "Libdl"]
+uuid = "c8ffd9c3-330d-5841-b78e-0817d7145fa1"
+version = "2.28.2+1"
+
+[[MozillaCACerts_jll]]
+uuid = "14a3606d-f60d-562e-9121-12d972cd8159"
+version = "2023.1.10"
+
+[[NetworkOptions]]
+uuid = "ca575930-c2e3-43a9-ace4-1e988b2c1908"
+version = "1.2.0"
+
+[[Pkg]]
+deps = ["Artifacts", "Dates", "Downloads", "FileWatching", "LibGit2", "Libdl", "Logging", "Markdown", "Printf", "REPL", "Random", "SHA", "Serialization", "TOML", "Tar", "UUIDs", "p7zip_jll"]
+uuid = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
+version = "1.10.0"
+
+[[PlutoDevMacros]]
+deps = ["HypertextLiteral", "InteractiveUtils", "MacroTools", "Markdown", "Pkg", "Random", "TOML"]
+git-tree-sha1 = "6ce1d9f7c078b493812161349c48735dee275466"
+uuid = "a0499f29-c39b-4c5c-807c-88074221b949"
+version = "0.5.8"
+
+[[Printf]]
+deps = ["Unicode"]
+uuid = "de0858da-6303-5e67-8744-51eddeeeb8d7"
+
+[[REPL]]
+deps = ["InteractiveUtils", "Markdown", "Sockets", "Unicode"]
+uuid = "3fa0cd96-eef1-5676-8a61-b3b8758bbffb"
+
+[[Random]]
+deps = ["SHA"]
+uuid = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+
+[[SHA]]
+uuid = "ea8e919c-243c-51af-8825-aaa63cd721ce"
+version = "0.7.0"
+
+[[Serialization]]
+uuid = "9e88b42a-f829-5b0c-bbe9-9e923198166b"
+
+[[Sockets]]
+uuid = "6462fe0b-24de-5631-8697-dd941f90decc"
+
+[[TOML]]
+deps = ["Dates"]
+uuid = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
+version = "1.0.3"
+
+[[Tar]]
+deps = ["ArgTools", "SHA"]
+uuid = "a4e569a6-e804-4fa4-b0f3-eef7a1d5b13e"
+version = "1.10.0"
+
+[[Tricks]]
+git-tree-sha1 = "aadb748be58b492045b4f56166b5188aa63ce549"
+uuid = "410a4b4d-49e4-4fbc-ab6d-cb71b17b3775"
+version = "0.1.7"
+
+[[UUIDs]]
+deps = ["Random", "SHA"]
+uuid = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
+
+[[Unicode]]
+uuid = "4ec0a83e-493e-50e2-b9ac-8f72acf5a8f5"
+
+[[Zlib_jll]]
+deps = ["Libdl"]
+uuid = "83775a58-1f1d-513f-b197-d71354ab007a"
+version = "1.2.13+1"
+
+[[nghttp2_jll]]
+deps = ["Artifacts", "Libdl"]
+uuid = "8e850ede-7688-5339-a07c-302acd2aaf8d"
+version = "1.52.0+1"
+
+[[p7zip_jll]]
+deps = ["Artifacts", "Libdl"]
+uuid = "3f19e933-33d8-53b3-aaab-bd5110c3b7a0"
+version = "17.4.0+2"
+"""
+
+# ╔═╡ Cell order:
+# ╠═cfb4d354-e0b2-4a04-a559-4fb88df33954
+# ╠═57c51c71-fd8d-440d-8262-9cccd1617c08
+# ╠═d9762fc1-fa2c-4315-a360-cc1cd9d70055
+# ╠═cbeff2f3-3ffe-4aaa-8d4c-43c44ee6d59f
+# ╠═ecfdee44-3e6c-4e7e-a039-1f7d05a875f8
+# ╠═9a90e5be-6c38-440d-a6d6-99dc25be5727
+# ╠═ca6e57a2-667f-4a77-a1ea-5f42192056d4
+# ╠═81c03bad-aed6-4b45-b3b5-edbba30c0bdb
+# ╠═ac83e4d6-1637-4646-865c-7d21ce010374
+# ╠═09458943-15a4-4251-82a8-e43299db53b5
+# ╠═d171e8f9-c939-4747-a748-5568ae4a4064
+# ╠═9dfb5236-9475-4bf6-990a-19f8f5519003
+# ╠═245fad1a-2f1e-4776-b048-6873e8c33f3b
+# ╠═456a3579-3c33-496c-bbf2-6e6e0d0ff102
+# ╠═8f7d7dac-d5e6-43f4-88e6-21179c78c3ef
+# ╟─00000000-0000-0000-0000-000000000001
+# ╟─00000000-0000-0000-0000-000000000002
