+++ test/runtests.jl
@@ -1,3 +1,4 @@
 using SafeTestsets
 
+@safetestset "Basic Widgets" begin include("basics.jl") end
 @safetestset "Extended Toc" begin include("toc.jl") end
